// src/__generated__/sendMessage.ts
/* tslint:disable */
/* eslint-disable */
// @generated
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL mutation operation: sendMessage
// ====================================================

export interface sendMessage_sendMessage {
  __typename: "SendMessageMutationResponse";
  ok: boolean;
  id: number | null;
}

export interface sendMessage {
  sendMessage: sendMessage_sendMessage;
}

export interface sendMessageVariables {
  payload: string;
  roomId?: number | null;
  userId?: number | null;
}